second commit for product stock reduction

+++ order-service/src/main/java/com/example/order_service/entity/OrderItems.java
@@ -32,7 +32,7 @@ public class OrderItems {
     private Double productPrice;
 
     @Column(nullable = false)
-    private Double productQuantity;
+    private Integer productQuantity;
 
     @Column(nullable = false)
     private Double totalAmount;
